pyproject.toml
[tool.poetry]
name = "team-utils"
version = "0.1.0"
description = "Team Utils for SEEK Dev Team"
authors = ["thesouldev <sasidharanofficial@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=4.2,<5"
djangorestframework = ">=3.15,<4"
psycopg2-binary = ">=2.9,<3"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = ">=5.3.1,<6"
django-health-check = ">=3.18.3,<4"
sendgrid = "^6.11.0"
django-cors-headers = "^4.4.0"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"
django-extensions = "^3.2.3"
django-stubs = "^5.1.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
isort = "^5.13.2"
black = "^24.4.2"
pyre-check = "^0.9.23"
coverage = "^7.6.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
skip-string-normalization = false

[tool.isort]
profile = "black"

[tool.coverage.run]
omit = ["*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError"
]

[tool.coverage.html]
directory = "htmlcov"
